Add option management to bin scripts

+++ bin/rekey_directory.sh
@@ -1,26 +1,130 @@
 #!/usr/bin/env bash
 
+# global definitions
 KO=1
 OK=0
+TRUE=0
+FALSE=1
 
-VAULT_PASS_ID=$1
-REKEY_DIR="inventory/${VAULT_PASS_ID}"
-VAULT_PASS_FILE="$HOME/.ansible_vault/${VAULT_PASS_ID}"
-VAULT_PASS_DIR="$(dirname ${VAULT_PASS_FILE})"
-VAULT_PASS_BASENAME="$(basename ${VAULT_PASS_FILE})"
-VAULT_VERIFY=0
-TMPROOT=temp
+function help {
+    echo "$0 OPTIONS <playbook> [ <playbook> ... ]"
+    echo
+    echo "OPTIONS:"
+    echo "   --vault-id   <vault>  # vault-id for key/rekey"
+    echo "  [--vault-dir] <dir>    # vault password file location"
+    echo "  [--generate]           # generate new vault password"
+    echo "  [--verify]             # restrict rekey to matching vault-ids"
+    echo "  [--debug]"
+}
 
-genxkpass() {
+function debug {
+    if [ "${DEBUG}" -eq "${TRUE}" ]; then
+        echo "$@"
+    fi
+}
+
+function check_requirement {
+    cmd=$1
+    command -v "${cmd}" >/dev/null 2>&1 || {
+        echo "${cmd} not found, aborting"
+        exit "${ERROR}"
+    }
+}
+
+function genxkpass() {
     curl -s 'https://xkpasswd.net/s/index.cgi' -H 'Content-Type: application/x-www-form-urlencoded; charset=UTF-8' --data 'a=genpw&n=1&c=%7B%22num_words%22%3A4%2C%22word_length_min%22%3A4%2C%22word_length_max%22%3A8%2C%22case_transform%22%3A%22RANDOM%22%2C%22separator_character%22%3A%22-%22%2C%22padding_digits_before%22%3A2%2C%22padding_digits_after%22%3A2%2C%22padding_type%22%3A%22NONE%22%2C%22random_increment%22%3A%22AUTO%22%7D' | jq '.passwords[0]' -r
 }
 
-TMPDIR=$(mktemp -d --tmpdir=${TMPROOT})
-REKEY_FILES=$(find "${REKEY_DIR}" -name "*.yml" -type f)
+check_requirement ansible-vault
+check_requirement yq
+check_requirement curl
 
-TMPVAULT="${TMPDIR}/vaults/${VAULT_PASS_BASENAME}"
+# set default values
+REKEY_FORCE=${FALSE}
+VAULT_VERIFY=${FALSE}
+TMPROOT=temp
+TMPDIR=$(mktemp -d --tmpdir=${TMPROOT})
+
+# parse options (https://stackoverflow.com/questions/192249/how-do-i-parse-command-line-arguments-in-bash)
+POSITIONAL=()
+while [[ $# -gt 0 ]]
+do
+    key="$1"
+
+    case $key in
+        --vault-id)
+            VAULT_PASS_ID="$2"
+            shift # past argument
+            shift # past value
+            ;;
+        --vault-dir)
+            ANSIBLE_VAULT_IDENTITY_DIR="$2"
+            shift # past argument
+            shift # past value
+            ;;
+        --rekey)
+            REKEY_FORCE=${TRUE}
+            shift # past argument
+            shift # past value
+            ;;
+        --verify)
+            VAULT_VERIFY=${TRUE}
+            shift # past argument
+            shift # past value
+            ;;
+        --help)
+            help
+            exit ${SUCCESS}
+            ;;
+        --debug)
+            DEBUG=${TRUE}
+            AWXCLI_VERBOSE="--verbose"
+            shift # past argument
+            ;;
+        *)  # unknown option
+            POSITIONAL+=("$1") # save it in an array for later
+            shift # past argument
+            ;;
+    esac
+done
+set -- "${POSITIONAL[@]}" # restore positional parameters
+
+# validate options
+if [ -z "${ANSIBLE_VAULT_IDENTITY_DIR}" ]; then
+    ANSIBLE_VAULT_IDENTITY_DIR=".ansible_vault"
+fi
+if [ -z "${VAULT_PASS_ID}" ]; then
+    echo "--vault-id <vault id> option is required"
+    exit ${KO}
+fi
+if [ ${#POSITIONAL[@]} -gt 0 ]; then
+    echo "Unknown positional arguments ${POSITIONAL[@]}"
+    exit ${KO}
+fi
+
+# set derived values
+REKEY_DIR="inventory/${VAULT_PASS_ID}"
+VAULT_PASS_FILE="$HOME/${ANSIBLE_VAULT_IDENTITY_DIR}/${VAULT_PASS_ID}"
+REKEY_FILES=$(find "${REKEY_DIR}" -name "*.yml" -type f)
+TMPVAULT="${TMPDIR}/vaults/$(basename ${VAULT_PASS_FILE})"
+
+if [ ! -d "$(dirname ${VAULT_PASS_FILE})" ]; then
+    mkdir "$(dirname ${VAULT_PASS_FILE})"
+    debug "Created vault password file directory $(dirname ${VAULT_PASS_FILE})"
+fi
+
+debug "Generating temp vault password file"
+
 mkdir "$(dirname ${TMPVAULT})"
-echo "$(genxkpass)" > "${TMPVAULT}"
+if [ ${REKEY_FORCE} -eq ${TRUE} ]; then
+    debug "Generating new vault password"
+    echo "$(genxkpass)" > "${TMPVAULT}"
+else
+    debug "cp ${VAULT_PASS_FILE}" "${TMPVAULT}"
+    cp "${VAULT_PASS_FILE}" "${TMPVAULT}"
+fi
+
+debug "Inspecting files [${REKEY_FILES}]"
 
 for file_name in $REKEY_FILES; do
 
@@ -30,6 +134,7 @@ for file_name in $REKEY_FILES; do
 
         for var_name in ${REKEY_VARS}; do
 
+            debug "Processing ${file_name}:${var_name}"
             TMPFILE="${TMPDIR}/$(basename "${file_name}")"
             encrypted=$(yq r "${file_name}" "${var_name}")
             if [ $? -ne 0 ]; then
@@ -39,30 +144,20 @@ for file_name in $REKEY_FILES; do
 
             if ! [[ "$encrypted" =~ "^\$ANSIBLE_VAULT;([^;]+;){2};${VAULT_PASS_ID}\n.*$" ]]; then
 
-                if [ "${VAULT_VERIFY}" -eq 1 ]; then
+                if [ ${VAULT_VERIFY} -eq ${TRUE}  ]; then
                     echo "Ignoring ${file_name}:${var_name} with mismatched vault id"
                     continue
                 fi
 
             fi
 
-            decrypt_success=1
-            for vault_pass_file in ${VAULT_PASS_FILE} $(find "${VAULT_PASS_DIR}/" -name ".${VAULT_PASS_BASENAME}*" -type f); do
-
-                decrypted=$(echo "${encrypted}" | ansible-vault decrypt --vault-id "${VAULT_PASS_ID}@${vault_pass_file}" 2>/dev/null)
-                if [ $? -ne 0 ]; then
-                    continue;
-                else
-                    decrypt_success=0
-                fi
-
-            done
-            if [ $decrypt_success -ne 0 ]; then
-                echo "error decrypting secret ${var_name} from files ${file_name}*"
+            decrypted=$(echo "${encrypted}" | ansible-vault decrypt --vault-id "${VAULT_PASS_ID}@${VAULT_PASS_FILE}")
+            if [ $? -ne 0 ]; then
+                echo "error decrypting secret ${var_name} from file ${file_name}"
                 exit "${KO}"
             fi
 
-            recrypted=$(echo "${decrypted}" | ansible-vault encrypt_string --encrypt-vault-id "${VAULT_PASS_ID}" --vault-password-file "${TMPVAULT}")
+            recrypted=$(echo "${decrypted}" | ansible-vault encrypt_string --encrypt-vault-id "${VAULT_PASS_ID}" --vault-password-file ${TMPVAULT})
             if [ $? -ne 0 ]; then
                 echo "error encrypting secret ${var_name} from file ${file_name}"
                 exit "${KO}"
@@ -71,7 +166,6 @@ for file_name in $REKEY_FILES; do
             echo "---" > "${TMPFILE}"
             echo "${var_name}: ${recrypted}" >> "${TMPFILE}"
             yq m -x -i "${file_name}" "${TMPFILE}" -I 2
-            echo -e "---\n$(cat ${file_name})" > "${file_name}"
 
         done
 
@@ -79,7 +173,11 @@ for file_name in $REKEY_FILES; do
 
 done
 
-mv "${VAULT_PASS_FILE}" "${VAULT_PASS_DIR}/.${VAULT_PASS_BASENAME}.$(date +%Y%m%d%H%M%S)"
-mv "${TMPVAULT}" "${VAULT_PASS_FILE}"
+if [ ${REKEY_FORCE} -eq ${TRUE} ]; then
+    if [ -f "${VAULT_PASS_FILE}" ]; then
+        mv "${VAULT_PASS_FILE}" "${VAULT_PASS_FILE}.$(date +%Y%m%d%H%M%S)"
+    fi
+    mv "${TMPVAULT}" "${VAULT_PASS_FILE}"
+fi
 
 rm -rf ${TMPDIR}
